chapter_009/src/main/java/ru/job4j/io/LogFilter.java
package ru.job4j.io;


import java.io.*;
import java.util.ArrayList;
import java.util.List;

/**
 * Created by Intellij IDEA.
 * User: Vitaly Zubov.
 * Email: Zubov.VP@yandex.ru.
 * Version: $Id$.
 * Date: 19.01.2021.
 */
public class LogFilter {
    public static List<String> filter(String file) {
        List<String> result = new ArrayList<>();
        try (BufferedReader bf = new BufferedReader(new FileReader(file))) {
            bf.lines().filter(e -> e.contains("404")).forEach(result::add);
        } catch (IOException e) {
            e.printStackTrace();
        }
        return result;
    }

    public static void save(List<String> log, String file) {
        try (PrintWriter pw = new PrintWriter(new BufferedOutputStream(new FileOutputStream(file)))) {
            log.forEach(e -> pw.write(e + "\n"));
        } catch (FileNotFoundException e) {
            e.printStackTrace();
        }
    }

    public static void main(String[] args) {
        List<String> log = filter("chapter_009/log.txt");
        save(log, "chapter_009/404.txt");
    }
}
